+++ xaib/utils/__init__.py
@@ -1,7 +1,4 @@
 from .utils import (
-    Filter,
-    KeyedComposer,
-    SimpleDataloader,
     batch_count_eq,
     batch_gini,
     batch_rmse,
@@ -9,3 +6,5 @@ from .utils import (
     minmax_normalize,
 )
 from .cache import ModelCache
+from .datasets import Filter, KeyedComposer, SimpleDataloader, NoiseApplier
+from .models import RandomBaseline, KNeighborsTransformer, RandomNeighborsBaseline
